requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
asyncpg==0.29.0
redis==5.0.1
aioredis==2.0.1

# LLM Support
openai==1.12.0
anthropic==0.18.1

# Optional: Observability (uncomment to enable)
# langfuse==2.50.0

# Optional: Additional LLM providers (uncomment if needed)
# boto3==1.34.0  # For AWS Bedrock
# azure-identity==1.15.0  # For Azure OpenAI

# API Features
slowapi==0.1.9
websockets==12.0
python-multipart==0.0.6

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.2

# Testing & Code Quality
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
ruff==0.1.9

# Utilities
python-dateutil==2.8.2

# MCP Server Support
mcp==1.0.0